feat: configure git branch
This enables users to add a branch as input to this action. When passed this branch will be used when updating the versions in maven.

This change gives users more freedom over their release process. For instance, if I want to support multiple versions and have different branches for managing this.

Imagine the following branches:
1. main
2. stable/1.0
3. stable/1.1

When I want to do a patch release on version 1.0, I want the versions to be updated on the stable/1.0 branch. I'd set the `branch` input to `stable/1.0` to achieve this.
Without this change making a new patch release for version 1.0 will result in an update of the main branch. Where the versions would've been 1.1 before, they'd be updated to 1.0 because of this patch.