fix: prevent second deployment to Maven Central
The RELEASE_PROFILE defaults to community-action-maven-release. Combined with community-hub-release-parent pom this will cause a deployment referred to as the `default-deployment`.
The CENTRAL_RELEASE_PROFILE also causes a deployment. This one is referred to as the `deploy-to-maven-central` deployment.

By default, the `default-deployment` will deploy to the Camunda Repository. However, because  in this command the `nexusUrl` gets overridden with `inputs.maven-url` this deployment will be done to the same repository that is used for the `deploy-to-maven-central` deployment.

Since the deployment to the Camunda Repository already happens in a separate command a few lines above this one, we don't need this profile at all.